Checked if system is already loaded, and beautified json

+++ services/src/main/kotlin/org/dxworks/inspectorgit/services/impl/LocalSystemsService.kt
@@ -51,6 +51,9 @@ class LocalSystemsService(private val loadedSystem: LoadedSystem,
     }
 
     fun load(id: String) {
+        if (id == loadedSystem.id)
+            return
+
         val systemFolder = getSystemFolder(id)
         if (!systemFolder.exists()) {
             localSystemRepository.deleteBySystemId(id)
